+++ src/addresses/mod.rs
@@ -1,4 +1,3 @@
-mod components;
 pub mod functions;
 pub mod models;
 pub mod page;
